# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Destiny-Ghost API",
    "version": "2.12.0",
    "description": "A Node Express application for receiving SMS/MMS Notifications around changes to the vendor wares in Bungie's Destiny and make ad-hoc queries in the database.",
    "license": {
      "name": "MIT",
      "url": "https://choosealicense.com/licenses/mit"
    }
  },
  "servers": [
    {
      "url": "https://api2.destiny-ghost.com",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bungieOAuth": {
        "type": "oauth2",
        "description": "This API uses OAuth 2.0 with Bungie.net as the Authorization Server. Clients should obtain a Bearer token from Bungie.net and include it in the Authorization header.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://www.bungie.net/en/OAuth/Authorize",
            "tokenUrl": "https://www.bungie.net/platform/app/oauth/token/",
            "scopes": {}
          }
        }
      },
      "authorizationKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Destiny-Ghost-Authorization",
        "description": "Requires administrative privileges. Specific header provided out-of-band."
      }
    },
...
  },
...
  "paths": {
...
    "/destiny/manifest": {
      "get": {
        "summary": "Get details about the latest Destiny manifest definition.",
        "operationId": "getDestinyManifest",
        "tags": [
          "Destiny"
        ],
        "parameters": [
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "description": "Return not modified if the manifest has not changed since the date and time provided.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the Destiny Manifest definition.",
            "headers": {
              "Last-Modified": {
                "description": "The date and time the manifest was last modified.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "description": "The cache control header.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified",
            "headers": {
              "Last-Modified": {
                "description": "The date and time the manifest was last modified.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "description": "The cache control header.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Download the latest Destiny manifest if the local copy is outdated.",
        "operationId": "downloadDestinyManifest",
        "tags": [
          "Destiny"
        ],
        "security": [
          {
            "authorizationKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the Destiny Manifest definition."
          }
        }
      }
    },
    "/destiny2/characters": {
      "get": {
        "summary": "Get a list of the user's characters.",
        "operationId": "getCharacters",
        "tags": [
          "Destiny 2"
        ],
        "security": [
          {
            "bungieOAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the current user's list of characters."
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/destiny2/inventory": {
      "get": {
        "summary": "Get the complete inventory of items.",
        "operationId": "getInventory",
        "tags": [
          "Destiny 2"
        ],
        "security": [
          {
            "authorizationKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the complete Destiny 2 item inventory."
          }
        }
      }
    },
    "/destiny2/manifest": {
      "get": {
        "summary": "Get details about the latest Destiny 2 manifest definition.",
        "operationId": "getDestiny2Manifest",
        "tags": [
          "Destiny 2"
        ],
        "parameters": [
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "description": "Return not modified if the manifest has not changed since the date and time provided.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the Destiny Manifest definition."
          }
        }
      },
      "post": {
        "summary": "Download the latest Destiny2 manifest if the local copy is outdated.",
        "operationId": "downloadDestiny2Manifest",
        "tags": [
          "Destiny 2"
        ],
        "security": [
          {
            "authorizationKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the Destiny Manifest definition."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/destiny2/xur": {
      "get": {
        "summary": "Get Xur's inventory if available.",
        "operationId": "getXurInventory",
        "tags": [
          "Destiny 2"
        ],
        "security": [
          {
            "bungieOAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns Xur's inventory."
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Xur could not be found."
          }
        }
      }
    },
...
  }
}